some refactors and more
