+++ packages/cli/src/util/git/connect-git-provider.ts
@@ -4,7 +4,7 @@ import chalk from 'chalk';
 import link from '../output/link';
 import { isAPIError } from '../errors-ts';
 import type { Dictionary } from '@vercel/client';
-import type { Org } from '@vercel-internals/types';
+import type { Org, Project, ProjectLinkData } from '@vercel-internals/types';
 import output from '../../output-manager';
 import list, { type ListChoice } from '../input/list';
 
@@ -144,4 +144,140 @@ export async function selectRemoteUrl(
     message: 'Which remote do you want to connect?',
     choices,
   });
+}
+
+export async function selectAndParseRemoteUrl(
+  client: Client,
+  remoteUrls: Dictionary<string>
+): Promise<RepoInfo | null> {
+  let remoteUrl: string;
+
+  if (Object.keys(remoteUrls).length > 1) {
+    output.log('Found multiple remote URLs.');
+    remoteUrl = await selectRemoteUrl(client, remoteUrls);
+  } else {
+    // If only one is found, get it — usually "origin"
+    remoteUrl = Object.values(remoteUrls)[0];
+  }
+
+  if (remoteUrl === '') {
+    output.log('Canceled');
+    return null;
+  }
+
+  const repoInfo = parseRepoUrl(remoteUrl);
+  if (!repoInfo) {
+    output.error(
+      `Failed to parse Git repo data from the following remote URL: ${link(
+        remoteUrl
+      )}`
+    );
+    return null;
+  }
+
+  return repoInfo;
+}
+
+export interface GitRepoCheckParams {
+  client: Client;
+  confirm: boolean;
+  gitProviderLink?: ProjectLinkData;
+  org: Org;
+  gitOrg: string;
+  project: Project;
+  provider: string;
+  repo: string;
+  repoPath: string;
+}
+
+export async function checkExistsAndConnect({
+  client,
+  confirm,
+  org,
+  project,
+  gitProviderLink,
+  provider,
+  repoPath,
+  gitOrg,
+  repo,
+}: GitRepoCheckParams) {
+  output.log(
+    `Connecting ${formatProvider(provider)} repository: ${chalk.cyan(repoPath)}`
+  );
+
+  if (!gitProviderLink) {
+    const connect = await connectGitProvider(
+      client,
+      project.id,
+      provider,
+      repoPath
+    );
+    if (typeof connect === 'number') {
+      return connect;
+    }
+  } else {
+    const connectedProvider = gitProviderLink.type;
+    const connectedOrg = gitProviderLink.org;
+    const connectedRepo = gitProviderLink.repo;
+    const connectedRepoPath = `${connectedOrg}/${connectedRepo}`;
+
+    const isSameRepo =
+      connectedProvider === provider &&
+      connectedOrg === gitOrg &&
+      connectedRepo === repo;
+    if (isSameRepo) {
+      output.log(
+        `${chalk.cyan(connectedRepoPath)} is already connected to your project.`
+      );
+      return 1;
+    }
+
+    const shouldReplaceRepo = await confirmRepoConnect(
+      client,
+      confirm,
+      connectedProvider,
+      connectedRepoPath
+    );
+    if (!shouldReplaceRepo) {
+      return 0;
+    }
+
+    await disconnectGitProvider(client, org, project.id);
+    const connect = await connectGitProvider(
+      client,
+      project.id,
+      provider,
+      repoPath
+    );
+    if (typeof connect === 'number') {
+      return connect;
+    }
+  }
+
+  output.log(
+    `Connected ${formatProvider(provider)} repository ${chalk.cyan(repoPath)}!`
+  );
+}
+
+async function confirmRepoConnect(
+  client: Client,
+  yes: boolean,
+  connectedProvider: string,
+  connectedRepoPath: string
+) {
+  let shouldReplaceProject = yes;
+  if (!shouldReplaceProject) {
+    shouldReplaceProject = await client.input.confirm(
+      `Looks like you already have a ${formatProvider(
+        connectedProvider
+      )} repository connected: ${chalk.cyan(
+        connectedRepoPath
+      )}. Do you want to replace it?`,
+      true
+    );
+    if (!shouldReplaceProject) {
+      output.log('Canceled. Repo not connected.');
+    }
+  }
+  return shouldReplaceProject;
 }
